XWIKI-4735: Integration test executor should use more than 2 second as timeout
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/core/trunk@26087 f329d543-caf0-0310-9063-dda96c69346f

+++ xwiki-shared-tests/src/main/java/org/xwiki/test/XWikiExecutor.java
@@ -196,7 +196,7 @@ public class XWikiExecutor
             method.getParams().setParameter(HttpMethodParams.RETRY_HANDLER,
                 new DefaultHttpMethodRetryHandler(0, false));
             // Set a socket timeout to ensure the server has no chance of not ansering to our request...
-            method.getParams().setParameter(HttpMethodParams.SO_TIMEOUT, new Integer(2000));
+            method.getParams().setParameter(HttpMethodParams.SO_TIMEOUT, new Integer(10000));
             
             try {
                 // Execute the method.
